Refactors CardGrid and MenuDropdown
